+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - 20210406
   1. spinner: use 'size' to control width, height, etc.
   2. spinner: add circle and change style
+  3. change 'timer-add' to 'timer-input'
 - 20210401
   1. refine some code and style
 - 20210330